requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
django==5.0.7
django-tinymce==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typing-extensions==4.12.2